PaymentService/requirements.txt
alembic==1.10.2
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-filter @ git+https://github.com/carltongibson/django-filter.git@b3a27f4fa5357dbfa367c2f60871299fe5c7aebe
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.9.0
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.3
whitenoise==6.3.0
